app/requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
PyMySQL
pytest
python-dotenv